outsourced cmd folder in modeling
